refactor(swift): split SettingsView.swift into focused files
Extract VaultSettingsSection, ModelSettingsSection, BackendSettingsSection,
PermissionsSettingsSection, EnvironmentSettingsSection, UISettingsSection,
and SetupWindowController into separate files.
